requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
openai==1.14.3
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
email-validator==2.2.0
Werkzeug==3.0.3
SQLAlchemy==2.0.31
httpx==0.27.0